improve calculator

+++ apcsp-practice/topic-3-1/calculator.py
@@ -1,6 +1,6 @@
-money_avaliable =  int(input("How much money do you have: "))
+money_avaliable =  float(input("How much money do you have: "))
 cases_buying = int(input("How many cases are you buying: "))
-case_cost = int(input("How much do the cases cost: "))
+case_cost = float(input("How much do the cases cost: "))
 
 total_cost = cases_buying * case_cost
 money_left = money_avaliable - total_cost
